Implement Jurnal module enhancements with new print styles, search functionality, and extender hooks, alongside minor Partner module refinements.

+++ public/modules/partner/partnerContactEdit.mjs
@@ -5,11 +5,11 @@ import * as pageHelper from '/public/libs/webmodule/pagehelper.mjs'
 const Extender = Ext.extenderContact ?? Ext
 
 
-const CurrentState = {}
 const Crsl =  Context.Crsl
 const CurrentSectionId = Context.Sections.partnerContactEdit
 const CurrentSection = Crsl.Items[CurrentSectionId]
 const Source = Context.Source
+const CurrentState = {}
 
 const TitleWhenNew = 'New Contact'
 const TitleWhenView = 'View Contact'
@@ -99,8 +99,10 @@ export async function openSelectedData(self, params) {
 
 		
 
+		const suspended = self.Modules.partnerHeaderEdit.getCurrentState().Actions.edit.isSuspended()
+
+		CurrentState.editDisabled = suspended
 		CurrentState.currentOpenedId = id
-		
 		
 		// jika posisi header dalam keadaan unlock (bisa edit, perlu cek kondisi data, untuk menentukan bisa diedit atau tidak)
 		if (!CurrentState.headerFormLocked) {
@@ -567,11 +569,11 @@ async function btn_save_click(self, evt) {
 
 
 		// Extender Saving
-		// export async function partnerContactEdit_dataSaved(self, data, frm) {}
 		const fn_datasaved_name = 'partnerContactEdit_dataSaved'
 		const fn_datasaved = Extender[fn_datasaved_name]
 		if (typeof fn_datasaved === 'function') {
-			await fn_datasaved(self, data, frm)
+			// export async function partnerContactEdit_dataSaved(self, data, frm) {}
+			await fn_datasaved(self, result, frm)
 		}
 
 
@@ -625,11 +627,38 @@ async function btn_del_click(self, evt) {
 	console.log('delete data')
 	let mask = $fgta5.Modal.createMask()
 	try {
+
+		// Extender Deleting
+		// export async function partnerContactEdit_dataDeleting(self, id, args) {}
+		const args = { cancelDelete: false }
+		const fn_datadeleting_name = 'partnerContactEdit_dataDeleting'
+		const fn_datadeleting = Extender[fn_datadeleting_name]
+		if (typeof fn_datadeleting === 'function') {
+			await fn_datadeleting(self, idValue, args)
+		}
+
+		// batalkan save, jika ada request cancel
+		if (args.cancelDelete) {
+			console.log('delete is canceled')
+			return
+		}
+
 		const result = await deleteData(self, idValue)
 		
+		
+
+		// Extender Delete
+		// export async function partnerContactEdit_dataDeleted(self, data) {}
+		const fn_datadeleted_name = 'partnerContactEdit_dataDeleted'
+		const fn_datadeleted = Extender[fn_datadeleted_name]
+		if (typeof fn_datadeleted === 'function') {
+			await fn_datadeleted(self, result)
+		}
+
+
 		// hapus current row yang dipilih di list
 		self.Modules.partnerContactList.removeCurrentRow(self)
-		
+
 		// kembali ke list
 		self.Modules.partnerContactList.Section.show()
